Move determinism function to utils/generic
