+++ GANDLF/utils/generic.py
@@ -1,4 +1,5 @@
 import os, datetime, sys
+import random
 import numpy as np
 import torch
 import SimpleITK as sitk
@@ -180,3 +181,21 @@ def get_array_from_image_or_tensor(input_tensor_or_image):
         return input_tensor_or_image
     else:
         raise ValueError("Input must be a torch.Tensor or sitk.Image or np.ndarray")
+
+def set_determinism(seed=42):
+    """
+    This function controls the randomness of the program. It sets the seed for both torch and numpy.
+    Args:
+        seed (int, optional): Seed to set. Defaults to 42.
+    """
+    random.seed(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+    torch.use_deterministic_algorithms(True, warn_only = True)
+
+    if torch.cuda.is_available():
+        os.environ["CUBLAS_WORKSPACE_CONFIG"] = ":16:8"
+        torch.cuda.manual_seed(seed)
+        torch.cuda.manual_seed_all(seed)
+        torch.backends.cudnn.deterministic = True
+        torch.backends.cudnn.benchmark = False
