--- GANDLF/utils/set_determinism.py
@@ -1,15 +1,0 @@
-import os, torch, random
-import numpy as np
-
-def set_determinism(seed=42):
-    random.seed(seed)
-    np.random.seed(seed)
-    torch.manual_seed(seed)
-    torch.use_deterministic_algorithms(True, warn_only = True)
-
-    if torch.cuda.is_available():
-        os.environ["CUBLAS_WORKSPACE_CONFIG"] = ":16:8"
-        torch.cuda.manual_seed(seed)
-        torch.cuda.manual_seed_all(seed)
-        torch.backends.cudnn.deterministic = True
-        torch.backends.cudnn.benchmark = False
